# server/openapi.yaml
openapi: 3.0.0
info:
  title: Neko API Reference
  version: "3.0.0"
  description: |
    Neko uses WebRTC and WebSocket besides REST API to provide real-time video streaming, interactive collaboration, and remote desktop functionality.

    ### Core Technologies

    - **Video Streaming over WebRTC**
      - Neko leverages **WebRTC** for real-time, low-latency, high-quality video and audio streaming.
      - Keyboard and mouse events are transmitted over the same WebRTC connection for seamless interaction.

    - **Real-Time Signaling and Communication over WebSocket**
      - **WebSocket** is used for sSignaling in WebRTC (e.g., session establishment, ICE candidate exchange).  
      - Live chat, session updates, and keyboard/mouse fallback are handled over WebSocket.

    - **REST API for Everything Else**
      - **Authentication & Session Management** - Secure user authentication and session handling.  
      - **Room and User Management** - Creating, modifying, and controlling user access to rooms.  
      - **Desktop Controls** - Adjusting screen configurations, managing keyboard and mouse events.
      - **System Health & Metrics** - Retrieving system health checks and operational insights.  

    ### How It All Works Together

    1. **WebRTC** handles media streaming for **real-time video, audio and data**.
    2. **WebSocket** ensures a **persistent, bidirectional** connection for low-latency interactions.  
...

...
  license:
    name: Apache 2.0
...
servers:
...
paths:
...
  /api/room/control:
    get:
      tags:
        - room-control
      summary: Get Control Status
      description: Retrieve the current control status of the room.
      operationId: controlStatus
      responses:
        '200':
          description: Control status retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControlStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/room/control/request:
    post:
      tags:
        - room-control
      summary: Request Control
      description: Request control of the room.
      operationId: controlRequest
      responses:
        '204':
          description: Control requested successfully.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: There is already a host.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /api/room/control/release:
    post:
      tags:
        - room-control
      summary: Release Control
      description: Release control of the room.
      operationId: controlRelease
      responses:
        '204':
          description: Control released successfully.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: There is already a host.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /api/room/control/take:
    post:
      tags:
        - room-control
      summary: Take Control
      description: Take control of the room.
      operationId: controlTake
      responses:
        '204':
          description: Control taken successfully.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/room/control/give/{sessionId}:
    post:
      tags:
        - room-control
      summary: Give Control
      description: Give control of the room to a specific session.
      operationId: controlGive
      parameters:
        - in: path
          name: sessionId
          description: The identifier of the session.
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Control given successfully.
        '400':
          description: Target session is not allowed to host.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/room/control/reset:
    post:
      tags:
        - room-control
      summary: Reset Control
      description: Reset the control status of the room.
      operationId: controlReset
      responses:
        '204':
          description: Control reset successfully.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: NEKO_SESSION
      description: Authentication using a session cookie.
    BearerAuth:
      type: http
      scheme: bearer
      description: Authentication using a Bearer token.
    TokenAuth:
      type: apiKey
      in: query
      name: token
      description: Authentication using a token passed as a query parameter.
  responses:
    NotFound:
      description: The specified resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    Unauthorized:
      description: The request requires user authentication.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    Forbidden:
      description: The server understood the request, but refuses to authorize it.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
  schemas:
    ErrorMessage:
      type: object
      properties:
        code:
          type: integer
          description: HTTP status code.
        message:
          type: string
          description: Detailed error message.
        error_code:
          type: string
          description: Stable application or protocol error code.
...
    ControlStatus:
      type: object
      properties:
        has_host:
          type: boolean
          description: Indicates if there is a host currently.
        host_id:
          type: string
          description: The ID of the current host, if any.
...
security:
  - BearerAuth: []
  - CookieAuth: []
  - TokenAuth: []